Move react code into main package
