+++ client/src/main/java/replicant/react4j/React4j.gwt.xml
@@ -1,0 +1,6 @@
+<module>
+  <inherits name='replicant.Replicant'/>
+  <inherits name='react4j.arez.ReactArez'/>
+
+  <source path=''/>
+</module>
